move exceptions to a higher level, where they are needed
